requirements.txt
aniso8601==8.1.0
astroid==2.4.2
autoenv==1.0.0
autopep8==1.5.4
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
libgravatar==0.2.4
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
passlib==1.7.4
pycodestyle==2.6.0
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
